Remove bundle.{Seq,If,Defer,newPhase,logString}, switch to regular functions (#2390)
## Changes
- Instead of constructing chains of mutators and then executing them,
execute them directly.
- Remove functionality related to chain-building: Seq, If, Defer,
newPhase, logString.
- Phases become functions that apply the changes directly rather than
construct mutator chains that will be called later.
- Add a helper ApplySeq to call multiple mutators, use it where
Apply+Seq were used before.

This is intended to be a refactoring without functional changes, but
there are a few behaviour changes:
- Since defer() is used to call unlock instead of bundle.Defer()
unlocking will now happen even in case of panics.
- In --debug, the phase names are are still logged once before start of
the phase but each entry no longer has 'seq' or phase name in it.
- The message "Deployment complete!" was printed even if
terraform.Apply() mutator had an error. It no longer does that.

## Motivation

The use of the chains was necessary when mutators were returning a list
of other mutators instead of calling them directly. But that has since
been removed, so now the chain machinery have no purpose anymore.

Use of direct functions simplifies the logic and makes bugs more
apparent and easy to fix.

Other improvements that this unlocks:
- Simpler stacktraces/debugging (breakpoints).
- Use of functions with narrowly scoped API: instead of mutators that
receive full bundle config, we can use focused functions that only deal
with sections they care about prepareGitSettings(currentGitSection) ->
updatedGitSection. This makes the data flow more apparent.
- Parallel computations across mutators (within phase): launch
goroutines fetching data from APIs at the beggining, process them once
they are ready.

## Tests
Existing tests.

+++ bundle/phases/destroy.go
@@ -11,6 +11,7 @@ import (
 	"github.com/databricks/cli/bundle/deploy/terraform"
 
 	"github.com/databricks/cli/libs/cmdio"
+	"github.com/databricks/cli/libs/diag"
 
 	"github.com/databricks/cli/libs/log"
 	terraformlib "github.com/databricks/cli/libs/terraform"
@@ -77,42 +78,65 @@ func approvalForDestroy(ctx context.Context, b *bundle.Bundle) (bool, error) {
 	return approved, nil
 }
 
-// The destroy phase deletes artifacts and resources.
-func Destroy() bundle.Mutator {
+func destroyCore(ctx context.Context, b *bundle.Bundle) diag.Diagnostics {
 	// Core destructive mutators for destroy. These require informed user consent.
-	destroyCore := bundle.Seq(
+	diags := bundle.ApplySeq(ctx, b,
 		terraform.Apply(),
 		files.Delete(),
-		bundle.LogString("Destroy complete!"),
 	)
 
-	destroyMutator := bundle.Seq(
-		lock.Acquire(),
-		bundle.Defer(
-			bundle.Seq(
-				terraform.StatePull(),
-				terraform.Interpolate(),
-				terraform.Write(),
-				terraform.Plan(terraform.PlanGoal("destroy")),
-				bundle.If(
-					approvalForDestroy,
-					destroyCore,
-					bundle.LogString("Destroy cancelled!"),
-				),
-			),
-			lock.Release(lock.GoalDestroy),
-		),
-	)
+	if !diags.HasError() {
+		cmdio.LogString(ctx, "Destroy complete!")
+	}
 
-	return newPhase(
-		"destroy",
-		[]bundle.Mutator{
-			// Only run deploy mutator if root path exists.
-			bundle.If(
-				assertRootPathExists,
-				destroyMutator,
-				bundle.LogString("No active deployment found to destroy!"),
-			),
-		},
-	)
+	return diags
+}
+
+// The destroy phase deletes artifacts and resources.
+func Destroy(ctx context.Context, b *bundle.Bundle) (diags diag.Diagnostics) {
+	log.Info(ctx, "Phase: destroy")
+
+	ok, err := assertRootPathExists(ctx, b)
+	if err != nil {
+		return diag.FromErr(err)
+	}
+
+	if !ok {
+		cmdio.LogString(ctx, "No active deployment found to destroy!")
+		return diags
+	}
+
+	diags = diags.Extend(bundle.Apply(ctx, b, lock.Acquire()))
+	if diags.HasError() {
+		return diags
+	}
+
+	defer func() {
+		diags = diags.Extend(bundle.Apply(ctx, b, lock.Release(lock.GoalDestroy)))
+	}()
+
+	diags = diags.Extend(bundle.ApplySeq(ctx, b,
+		terraform.StatePull(),
+		terraform.Interpolate(),
+		terraform.Write(),
+		terraform.Plan(terraform.PlanGoal("destroy")),
+	))
+
+	if diags.HasError() {
+		return diags
+	}
+
+	hasApproval, err := approvalForDestroy(ctx, b)
+	if err != nil {
+		diags = diags.Extend(diag.FromErr(err))
+		return diags
+	}
+
+	if hasApproval {
+		diags = diags.Extend(destroyCore(ctx, b))
+	} else {
+		cmdio.LogString(ctx, "Destroy cancelled!")
+	}
+
+	return diags
 }
